Bug.MDD146 - Fixing Bugs in entrypoint_apache.sh for MISP-Server 2.4.97-2.4.99 (#22)
## Changelog for bug.MDD146 - Fixing Bugs in entrypoint_apache.sh for MISP-Server 2.4.97-2.4.9b
### Update Informations 
The entrypoint_apache.sh script takes care of the settings of apache2 and the basic settings of MISP itself. However, there were bugs regarding the baseurl. This was not set correctly. This caused the browsers to report an error after logging in. It said that the page was endlessly redirected.

### General Changes
No general changes were made.

### Fixes & Improvements
- Fixed an Bug in entrypoint_apache.sh script and in the Dockerfile

### Detailed Changes
- We have changed the order in the dockerfile for the git dependencies of MISP.
- We have changed the baseurl sed command to set it up in config.php and in the cake settings.
- We have changed the apache2 configuration.